autoload extractors/props, add some more extract tests
